requirements.txt
pyrogram
tgcrypto
python-dotenv
asyncpg
fastapi
uvicorn[standard]
httpx
Pillow
